Add files via upload
Successfully able to change the orientation of the solar panel around z-axis with manually defined angles on both -ve and +ve angles from the artificial 'N' which is estimated while calibrating. This code will eventually help to rotate the structure around z-axis which will align with the actual azimuth angle later.